day 11 part 2 was the most dofficult so far! had to alter many data types to permit larger integers

+++ day 11/day11.cpp
@@ -33,8 +33,8 @@ int main(int argc, char* argv[]) {
     Solver<Parts::PART1> part_1(data);
     Solver<Parts::PART2> part_2(data);
 
-    int part_1_solution = part_1.Solve();
-    int part_2_solution = part_2.Solve();
+    unsigned long part_1_solution = part_1.Solve();
+    unsigned long part_2_solution = part_2.Solve();
 
     std::cout << "Part 1: " << part_1_solution << "\n";
     std::cout << "Part 2: " << part_2_solution << "\n";
